move query batches into its own module
