+++ lib/orel.rb
@@ -29,10 +29,11 @@ require 'orel/finder'
 require 'orel/object'
 require 'orel/operator'
 require 'orel/options'
+require 'orel/query'
+require 'orel/query_batches'
 require 'orel/relation'
 require 'orel/schema_generator'
 require 'orel/simple_associations'
-require 'orel/query'
 require 'orel/table'
 require 'orel/validator'
