+++ lib/orel/query/select.rb
@@ -1,6 +1,8 @@
 module Orel
   class Query
     class Select
+      include Orel::QueryBatches
+
       attr_reader :projected_joins
       attr_reader :locked_for_query
       attr_reader :batch_size
@@ -65,21 +67,6 @@ module Orel
       def unlock_for_query!
         @locked_for_query = false
         nil
-      end
-
-      # Public: Specify that you want the results to be queried in batches.
-      #
-      # options - Hash of options.
-      #           :size  - Number of rows to query in each batch (default: 1000).
-      #           :group - Boolean whether to enumerate results individually or by batch.
-      #           :order - Boolean whether to order the query by the key, or leave to natural order.
-      #
-      # Returns nothing.
-      def query_batches(options)
-        @batch_size = options.delete(:size) || 1000
-        @batch_group = options.delete(:group) || false
-        @batch_order = options.key?(:order) ? options.delete(:order) : true
-        raise ArgumentError, "Unknown options: #{options.keys.inspect}" if options.any?
       end
 
       protected
